refactor(workflow-controller): separate state and task runtime
The workflow controller mixed process coordination, state transitions, and asynchronous task recovery in one file.

Extract pure WorkflowRun transitions and the character-template task lifecycle behind the existing controller facade.

Keep the public API and runtime behavior unchanged while making later workflow steps easier to add.
